career-mentor-agent/pyproject.toml
[project]
name = "career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.94.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
